Use ScopedValue for pipeline to ease casuality (#481)
* Use ScopedValue for pipeline to ease casuality

* Fix issues

* Move dependant setting

* Fix rebase issues

* Fix siize issues for test

+++ generator/src/main/java/com/faforever/neroxis/generator/terrain/LandBridgeTerrainGenerator.java
@@ -20,7 +20,7 @@ public class LandBridgeTerrainGenerator extends PathedTerrainGenerator {
     @Override
     protected void landSetup() {
         int mapSize = map.getSize();
-        float maxStepSize = mapSize / 128f;
+        int maxStepSize = mapSize / 128;
         int numPaths = 32 / generatorParameters.spawnCount();
 
         List<Vector2> team0Spawns = map.getSpawns()
@@ -45,7 +45,7 @@ public class LandBridgeTerrainGenerator extends PathedTerrainGenerator {
     @Override
     protected void plateausSetup() {
         int mapSize = map.getSize();
-        float maxStepSize = mapSize / 128f;
+        int maxStepSize = mapSize / 128;
         int maxMiddlePoints = 2;
         int numPaths = (int) (16 * plateauDensity) / symmetrySettings.spawnSymmetry().getNumSymPoints();
         int bound = mapSize / 4;
@@ -53,7 +53,7 @@ public class LandBridgeTerrainGenerator extends PathedTerrainGenerator {
 
         MapMaskMethods.pathInEdgeBounds(random.nextLong(), plateaus, maxStepSize, numPaths, maxMiddlePoints, bound,
                                         (float) (StrictMath.PI / 2));
-        plateaus.inflate(mapSize / 256f).setSize(mapSize / 4);
+        plateaus.inflate(mapSize / 256).setSize(mapSize / 4);
         plateaus.dilute(.5f, 4).setSize(mapSize + 1);
         plateaus.blur(12);
     }
